Fixed corr var scanner issue (now reads a double)
updated Target class to work with type change

+++ SRIM/src/Main.java
@@ -176,7 +176,7 @@ public class Main {
 				target = new Target(
 						new Scanner(density.getText()).nextDouble(),
 						state.getSelectedIndex(),
-						new Scanner(corr.getText()).nextInt());
+						new Scanner(corr.getText()).nextDouble());
 				makeElementFrame();
 			}  
 		});
